+++ railties/test/application/middleware_test.rb
@@ -56,8 +56,17 @@ module ApplicationTests
       assert !middleware.include?("Rack::Sendfile"), "Rack::Sendfile is not included in the default stack unless you set config.action_dispatch.x_sendfile_header"
     end
 
-    test "Rack::Cache is present when action_controller.perform_caching is set" do
+    test "Rack::Cache is not included by default" do
       add_to_config "config.action_controller.perform_caching = true"
+
+      boot!
+
+      assert !middleware.include?("Rack::Cache"), "Rack::Cache is not included in the default stack unless you set config.action_dispatch.rack_cache"
+    end
+
+    test "Rack::Cache is present when action_controller.perform_caching is set and action_dispatch.rack_cache is set" do
+      add_to_config "config.action_controller.perform_caching = true"
+      add_to_config "config.action_dispatch.rack_cache = true"
 
       boot!